Moved `dummy_plugin` to `examples/dummy_plugin` at project root
- Added a `README.md` and a `pyproject.toml` for it
- Added it as a dev dependency to nautobot's root `pyproject.toml`
- Asserted that unit tests ran:

```
Creating test database for alias 'default'...

System check identified no issues (0 silenced).
.....................
----------------------------------------------------------------------
Ran 21 tests in 0.134s

OK
Destroying test database for alias 'default'...
```
