+++ examples/README.md
@@ -1,0 +1,3 @@
+# Nautobot Examples
+
+This directory contains examples of extending Nautobot using plugins or other exciting features!
