+++ examples/dummy_plugin/dummy_plugin/admin.py
