+++ examples/dummy_plugin/dummy_plugin/datasources.py
