+++ examples/dummy_plugin/dummy_plugin/jobs.py
