+++ examples/dummy_plugin/dummy_plugin/navigation.py
