+++ examples/dummy_plugin/dummy_plugin/template_content.py
