+++ examples/dummy_plugin/dummy_plugin/urls.py
