+++ examples/dummy_plugin/dummy_plugin/views.py
