+++ pyproject.toml
@@ -95,6 +95,8 @@ black = "^20.8b1"
 coverage = "~5.4"
 # Tool for debugging Django
 django-debug-toolbar = "~3.2.1"
+# Nautobot dummy plugin used for testing
+dummy-plugin = {path = "examples/dummy_plugin", develop = true}
 # Code style checking and limited static analysis
 flake8 = "~3.9.1"
 # Alternative to Make, CLI based on `tasks.py`
